clients sending all at once

+++ p4/skeleton-2pc/run.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 
-rm -f /tmp/coordinator.log
-rm -f /tmp/participant_*.log
+if [ "$1" == "" ]; then
+    MODE=run
+    rm -f /tmp/coordinator.log
+    rm -f /tmp/participant_*.log
+else
+    MODE=check
+fi
 
-./target/debug/cs380p-2pc -s .5 -c 2 -p 2 -r 10 -v 5 -m run
+./target/debug/cs380p-2pc -s .8 -c 5 -p 5 -r 10 -v 5 -m $MODE
